44. Error handling: add exceptions

+++ 08-concepts/app/db/db_user.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm import Session
 from app.schemas import UserBase
 from app.db.models import DbUser
@@ -21,13 +22,18 @@ def get_all_user(db: Session):
 
 
 def get_user(db: Session, id: int):
-    # TODO: handle any exceptions
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
+    return user
 
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
-    # TODO: handle any exceptions
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
@@ -39,7 +45,9 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
-    # TODO: handle any exceptions
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
     db.delete(user)
     db.commit()
     return 'ok'
